Refactor: 캘린더와 카테고리 뷰 페이지 컴포넌트 구조 변경
